bigquery/requirements.txt
google-cloud-bigquery>=3.20.0
google-cloud-storage>=2.14.0
vertexai>=1.43.0
google-auth>=2.25.2
tenacity>=8.2.3
mcp[cli]>=1.0.0
pydantic>=2.0.0,<2.11.0
uvicorn[standard]>=0.27.0
starlette>=0.27.0
anyio>=4.0.0
python-dateutil>=2.8.0

